MCV achitecture being implemented, eventListeners still missing from view models
