+++ index.html
@@ -178,7 +178,6 @@
             <button
               type="button"
               class="btn btn-danger d-none btn--deleteTransaction"
-              data-bs-dismiss="modal"
             >
               Delete
             </button>
@@ -205,6 +204,6 @@
       }
     </script>
 
-    <script src="./script.js" type="module"></script>
+    <script src="./assets/js/index.js" type="module"></script>
   </body>
 </html>
